0.8.1.0: fix for aeson-1.0
This should now work with aeson-1.2.*. The support for
auto-building map instances instroduced in aeson-1.0
was interfering with the use of api-tools 'with'
constructions to achieve the same end/ To minimise the disruption
the 'with' constructions for maps were abandoned and all parts
of the schema containing maps were rebuilt with in Haskell
(without recourse to the api-tools DSL), with the old schema
renamed under renamed types. The stores are read with the
original schem to maintain identical JSON representation and
converted into the original internal format (with the maps)
using conversion functions, avoiding internal disruption.

Some adjustments in the way time formats %Q have been
eliminated by applying cleanup filter to hit the historical
format. This should ensure that existing keystore
signatures remain valid after upgrading to this keystore
without the need to re-sign.